domain: move system structs to types sub-package
