+++ pkg/bindings/system/system.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/containers/podman/v4/libpod/define"
 	"github.com/containers/podman/v4/pkg/bindings"
-	"github.com/containers/podman/v4/pkg/domain/entities"
 	"github.com/containers/podman/v4/pkg/domain/entities/types"
 	"github.com/sirupsen/logrus"
 )
@@ -63,9 +62,9 @@ func Events(ctx context.Context, eventChan chan types.Event, cancelChan chan boo
 }
 
 // Prune removes all unused system data.
-func Prune(ctx context.Context, options *PruneOptions) (*entities.SystemPruneReport, error) {
+func Prune(ctx context.Context, options *PruneOptions) (*types.SystemPruneReport, error) {
 	var (
-		report entities.SystemPruneReport
+		report types.SystemPruneReport
 	)
 	conn, err := bindings.GetClient(ctx)
 	if err != nil {
@@ -84,10 +83,10 @@ func Prune(ctx context.Context, options *PruneOptions) (*entities.SystemPruneRep
 	return &report, response.Process(&report)
 }
 
-func Version(ctx context.Context, options *VersionOptions) (*entities.SystemVersionReport, error) {
+func Version(ctx context.Context, options *VersionOptions) (*types.SystemVersionReport, error) {
 	var (
-		component entities.ComponentVersion
-		report    entities.SystemVersionReport
+		component types.SystemComponentVersion
+		report    types.SystemVersionReport
 	)
 	if options == nil {
 		options = new(VersionOptions)
@@ -135,8 +134,8 @@ func Version(ctx context.Context, options *VersionOptions) (*entities.SystemVers
 
 // DiskUsage returns information about image, container, and volume disk
 // consumption
-func DiskUsage(ctx context.Context, options *DiskOptions) (*entities.SystemDfReport, error) {
-	var report entities.SystemDfReport
+func DiskUsage(ctx context.Context, options *DiskOptions) (*types.SystemDfReport, error) {
+	var report types.SystemDfReport
 	if options == nil {
 		options = new(DiskOptions)
 	}
