+++ pkg/domain/entities/types/pods.go
@@ -1,0 +1,6 @@
+package types
+
+type PodPruneReport struct {
+	Err error
+	Id  string //nolint:revive,stylecheck
+}
